Improve loading and access to environment variables
- Make sure that server environment variables cannot be accessed or accidentally leaked by importing a module in client components.
- Only do the schema validation once.
- Make sure the env objects cannot be changed with Object.freeze.
