+++ src/app/faro/faro/faro.ts
@@ -6,8 +6,8 @@ import {
   initializeFaro,
 } from "@grafana/faro-web-sdk";
 import { TracingInstrumentation } from "@grafana/faro-web-tracing";
-
-import { isLocalOrDemo, publicEnv } from "@/constants/envs";
+import { isLocalOrDemo } from "@/env-variables/envHelpers";
+import { publicEnv } from "@/env-variables/publicEnv";
 
 export const initFaro = (): Faro | null => {
   if (
