+++ src/auth/redirectToLogin.ts
@@ -1,9 +1,9 @@
-import { publicEnv } from "@/constants/envs";
+import { publicEnv } from "@/env-variables/publicEnv";
 import { redirect } from "next/navigation";
 
 export const redirectToLogin = (narmestelederid: string) => {
   const loginPath = `/oauth2/login/redirect?redirect=${encodeURIComponent(
-    `${publicEnv.NEXT_PUBLIC_BASE_PATH}/oppfolgingsplan/${narmestelederid}`,
+    `${publicEnv.NEXT_PUBLIC_BASE_PATH}/oppfolgingsplan/${narmestelederid}`
   )}`;
   return redirect(loginPath);
 };
